feat: split app creation code from server initialization code
